+++ src/main/java/com/hseapple/dao/UserCourseEntity.java
@@ -1,9 +1,12 @@
 package com.hseapple.dao;
-import com.fasterxml.jackson.annotation.JsonFormat;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
-import java.sql.Timestamp;
 
 @Entity
 @Table(name = "user_course", schema = "public")
@@ -16,11 +19,11 @@ public class UserCourseEntity {
         this.id = id;
     }
 
-    public Long getCourseID() {
+    public Integer getCourseID() {
         return courseID;
     }
 
-    public void setCourseID(Long courseID) {
+    public void setCourseID(Integer courseID) {
         this.courseID = courseID;
     }
 
@@ -32,11 +35,11 @@ public class UserCourseEntity {
         this.userID = userID;
     }
 
-    public Long getRoleID() {
+    public Integer getRoleID() {
         return roleID;
     }
 
-    public void setRoleID(Long roleID) {
+    public void setRoleID(Integer roleID) {
         this.roleID = roleID;
     }
 
@@ -46,7 +49,7 @@ public class UserCourseEntity {
 
     @NotNull
     @Column(name = "courseid")
-    private Long courseID;
+    private Integer courseID;
 
     @NotNull
     @Column(name = "userid")
@@ -54,7 +57,7 @@ public class UserCourseEntity {
 
     @NotNull
     @Column(name = "roleid")
-    private Long roleID;
+    private Integer roleID;
